agent/requirements.txt
google-adk
google-cloud-firestore
opentelemetry-instrumentation-google-genai